Added Teacher data table and implemented TeacherController in backend

+++ server/API/Controllers/StatisticalController.cs
@@ -48,7 +48,7 @@ namespace SchoolBook.API.Controllers
         
         /*All schools statistics (Admin panel)*/
         [HttpGet("scores")]
-        public int GetSchoolScore()
+        public double GetSchoolScore()
         {
             return StatisticalService.SchoolAverageScore();
         }
